Add a basic data layer for interacting with mongo
This is certainly overkill for this size of project, but after dealing
with a couple of very complex and messy data layers on the
last few apps I worked on, I was curious to try building something
simpler that still gets the job done. I'm fairly happy with how this
turned out! Given how simple it is, I feel like the resulting API
is quite ergonomic.

Obviously, it would need some low-level tests if was ever going to
go into production. I'd also be curious to try sprinkling in some
generic types, but I'm not sure it would really be worth the
effort.